Ignore Authorization Basic header
eve-auth-jwt raises NotImplementedError when an Authorization Basic header is
present. This will happen if you are using eve-swagger, and the swagger docs
are protected with Auth Basic, and the Bearer token is not yet defined. In that
case, the browser sends the Authorization Basic header, which makes eve-auth-jwt
try to call check_auth, which is not implemented.

We are changing the authorized so it will ignore that header.